pyproject.toml
[tool.poetry]
name = "comfy-executors"
version = "0.1.0"
description = ""
authors = ["Kristian Klemon <kristian.klemon@gmail.com>"]
readme = "README.md"
packages = [
    { include = "comfy_executors", from = "." }
]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.3.0"
requests = "^2.31.0"
runpod = "^1.6.2"
jinja2 = "^3.1.3"
comfy-api-client = {git = "https://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@github.com/Infinishot/comfy-api-client.git"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
